Rename `components` to `elements`
