+++ packages/buttons/src/index.ts
@@ -5,9 +5,9 @@
  * found at http://www.apache.org/licenses/LICENSE-2.0.
  */
 
-export { default as Button } from './components/Button';
-export { default as Anchor } from './components/Anchor';
-export { default as ButtonGroup } from './components/ButtonGroup';
-export { default as ChevronButton } from './components/ChevronButton';
-export { default as IconButton } from './components/IconButton';
-export { default as SplitButton } from './components/SplitButton';
+export { default as Button } from './elements/Button';
+export { default as Anchor } from './elements/Anchor';
+export { default as ButtonGroup } from './elements/ButtonGroup';
+export { default as ChevronButton } from './elements/ChevronButton';
+export { default as IconButton } from './elements/IconButton';
+export { default as SplitButton } from './elements/SplitButton';
